More HP and Stamina stuff
regens stamina
loses stamina when jumping
loses stamina when running
loses hp by enter
loses hp by stay

+++ Assets/Script/Parts/Torso/BasicTorsoScriptable.cs
@@ -7,5 +7,5 @@ public class BasicTorsoScriptable : ScriptableObject
 {
     public int maxHp;
     public int maxStamina;
-    public int staminaRegen;
+    public float staminaRegen;
 }
